+++ qt_ui/windows/mission/flight/settings/QFlightStartType.py
@@ -1,4 +1,10 @@
-from PySide2.QtWidgets import QComboBox, QGroupBox, QHBoxLayout, QLabel
+from PySide2.QtWidgets import (
+    QComboBox,
+    QGroupBox,
+    QHBoxLayout,
+    QLabel,
+    QVBoxLayout,
+)
 
 from gen.flights.flight import Flight
 from qt_ui.models import PackageModel
@@ -11,8 +17,9 @@ class QFlightStartType(QGroupBox):
         self.package_model = package_model
         self.flight = flight
 
-        self.layout = QHBoxLayout()
-        self.start_type_label = QLabel("Start type : ")
+        self.layout = QVBoxLayout()
+        self.main_row = QHBoxLayout()
+        self.start_type_label = QLabel("Start type:")
         self.start_type = QComboBox()
 
         for i, st in enumerate([b for b in ["Cold", "Warm", "Runway", "In Flight"]]):
@@ -21,8 +28,14 @@ class QFlightStartType(QGroupBox):
                 self.start_type.setCurrentIndex(i)
 
         self.start_type.currentTextChanged.connect(self._on_start_type_selected)
-        self.layout.addWidget(self.start_type_label)
-        self.layout.addWidget(self.start_type)
+        self.main_row.addWidget(self.start_type_label)
+        self.main_row.addWidget(self.start_type)
+
+        self.layout.addLayout(self.main_row)
+        self.layout.addWidget(QLabel(
+            "Any option other than Cold will make this flight non-targetable " +
+            "by OCA/Aircraft missions. This will affect game balance."
+        ))
         self.setLayout(self.layout)
 
     def _on_start_type_selected(self):