+++ qt_ui/windows/mission/flight/settings/QGeneralFlightSettingsTab.py
@@ -2,7 +2,6 @@ from PySide2.QtCore import Signal
 from PySide2.QtWidgets import QFrame, QGridLayout, QVBoxLayout
 
 from game import Game
-from gen.ato import Package
 from gen.flights.flight import Flight
 from qt_ui.models import PackageModel
 from qt_ui.windows.mission.flight.settings.QFlightDepartureDisplay import \
@@ -24,21 +23,12 @@ class QGeneralFlightSettingsTab(QFrame):
         super().__init__()
 
         layout = QGridLayout()
-        flight_info = QFlightTypeTaskInfo(flight)
-        flight_departure = QFlightDepartureDisplay(package_model, flight)
-        flight_slots = QFlightSlotEditor(package_model, flight, game)
-        flight_start_type = QFlightStartType(package_model, flight)
-        flight_custom_name = QFlightCustomName(flight)
-        layout.addWidget(flight_info, 0, 0)
-        layout.addWidget(flight_departure, 1, 0)
-        layout.addWidget(flight_slots, 2, 0)
-        layout.addWidget(flight_start_type, 3, 0)
-        layout.addWidget(flight_custom_name, 4, 0)
+        layout.addWidget(QFlightTypeTaskInfo(flight), 0, 0)
+        layout.addWidget(QFlightDepartureDisplay(package_model, flight), 1, 0)
+        layout.addWidget(QFlightSlotEditor(package_model, flight, game), 2, 0)
+        layout.addWidget(QFlightStartType(package_model, flight), 3, 0)
+        layout.addWidget(QFlightCustomName(flight), 4, 0)
         vstretch = QVBoxLayout()
         vstretch.addStretch()
-        layout.addLayout(vstretch, 3, 0)
+        layout.addLayout(vstretch, 5, 0)
         self.setLayout(layout)
-
-        flight_start_type.setEnabled(flight.client_count > 0)
-        flight_slots.changed.connect(
-            lambda: flight_start_type.setEnabled(flight.client_count > 0))